Modern syntaxic language MSL
takes compile.txt and exports compile.exe. for now, msl.exe will copy its own PE header and paste it in compile.exe, then it will read and append the contents of compile.txt to compile.exe. Despite the function is simple, but the main logic for the compiling has been written.

There are 2 main buffers of size 0x1000, buffer1 for storing ASCII letters pasted from compile.txt and buffer2 for storing the compiled machine code that will then be copied into compile.exe. the main decode loop, which provides one ASCII letter per iteration, evaluates which syntaxic path to be taken: whether the read letter should be part of a dll import declaration, a variable declaration , a function declaration or an if statement; then the main decode loop jumps to specific decode instructions which will decide what to do with the ASCII byte. The specific decode instructions can specify a number in a register to force the main decode to pick the same decode path (no matter what the next character is) for the next character.

All the subsequent compiling logic has been planned: symbol table entries will have 16 bytes, 8 bytes to store the pointer to the variable name, 2 bytes to store the variable type and 6 remaining bytes to store the runtime pointer to the stack, which will contain the variable value. in x86-64, 64-bit addresses are sign-extended from the 47th bit to the 63th, so in order to use the runtime pointer, the beginning 2 bytes containing the variable type will need to be cleared or set depending on the 47th bit of the address. This design was chosen to minimize the space taken by each symbol entry and to maximize the number of symbols. The space reserved for the symbol table will be 0x4000 bytes, and each symbol entry is 16 byte, so the compiler will allow at most 1024 globa symbols ("global" as in symbols that were declared outside function or if statement scopes)