proof of work
after dll processing, there's address processing. in the language, you assign values like this:
newaddressname: 5+5
existingaddressname+ 5*2 - ...(address indexing): 57+23+[anotheraddress]
where ":" means "the address on the left contains the value on the right" and "[address]" means "the value located at that address"
in the upload, the code is now able to put those expressions in a table that I named the "arithmetic table" where each entry is 16 byte. here's the layout of an entry:
byte 0: following operator's priority
byte 1: following operator's identifier
byte 2: immediate type (if the entry represents an immediate)
quadword 1 (the second quadword): pointer to the variable in the symbol table that the arithmetic entry represents, OR immediate value
I'm not gonna say too much, but let's say you have this expression:
hello+5*09                    ; the extra 0 is not a typo
then the arithmetic table should look like this:
entry 1:
first quadword: 9 (priority of "+"), 14 (identifier of "+"), 0 (no type because the entry isn't an immediate), 0, 0, 0, 0, 0
second quadword: pointer to the symbol with the name "hello"
entry 2:
first quadword: 10 (priority of "*"), 16 (identifier of "*"), 1 (type is int64), 0, 0, 0, 0, 0
second quadword: 5
entry 3:
first quadword: 0, 0 (there's no following operator after 9), 2 (type is int32, because the extra "0" before the 9 means int32), 0, 0, 0, 0, 0
second quadword: 9
after this, another loop will read the arithmetic table and translate everything ORDERLY to machine code by checking the priority byte of each entry, nulling each entry as it's translated
precision for immediates:
9: int64
09: int32
009: uint16
0009: uint8
.9: float64
0.9: float32
this way, there's no bloaty type specifier before each number
if coding the translation of the arithmetic entry into machine code is easy, I might finish this compiler in one week