proof of work 2
this update (almost) allows new syntaxes:
1. name[32] for the declaration of "name" at rsp - 32. the substraction of rsp is done
2. name+24: 7 for the writing of the 64-bit value 7 at the address name+24, where name is a fixed stack location (no matter how the stack pointer moves, the compiler will be able to track the offset)
every assignment statement (second example), for whatever expression is written before the colons ":", that expression needs to start with a name (an address) so the compiler can recognize that it's an assignment statement
another update is that the arithmetic loop can now evaluate in what order to encode expressions:
in a+b*c, the compiler knows to encode * first, then +. this is not hardcoded stuff. it's a logical parser that assigns each operator (+, -, *, /, %, etc) a priority value. that priority value is increased inside an open parenthesis "(" and decreased after passing a closed parenthesis
here are the operators and their priorities
1. | logical or
2. & logical and
3. || bitwise or
4. |& bitwise xor
5. && bitwise and
6. =, ~ equal and unequal
7. <, >, <=, >= relation
8. <<, >> shift
9. +, -
10. *, / (floor division or float), %
11. -, ! unary negate and not
12. [ address dereferencing; "[name]" means "the value at address "name"", interpreted as a sticky type, aka the same type as the first operand [name] interacts with
+12 ( and [
-12 ) and ]
I don't want to tell how the arithmetic encoding logic works, but you can inspect the code to find the logic or ask me in person if you are an interviewer